+++ core/hashing.js
