+++ core/signature.js
